+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.7.1] - 2024-06-17
+### Fixed
+- Fixed `AggressiveScaleDown` model in GCP Ocean.
+
+## [3.7.0] - 2024-06-14
+### Added
+- Added `AggressiveScaleDown` model in GCP Ocean.
+
+## [3.6.0] - 2024-05-30
+### Added
+- Added `AmiAutoUpdate` model in Aws Ocean.
+
+## [3.5.0] - 2024-05-26
+### Added
+- Added `linux_o_s_config` field in the `NodePoolProperties` model
+ for Azure AKS Ocean.
+
+## [3.4.0] - 2024-05-24
+### Added
+- Added `should_deregister_from_lb` field in the `DeallocationConfig` model
+ for Azure Stateful Node.
+
+## [3.3.0] - 2024-05-22
+### Added
+- Added support for `Task` in Ocean AKS `Scheduling` model.
+
+## [3.2.0] - 2024-05-17
+### Added
+- Added support for `GpuType` in Ocean AKS `Filters` model.
+
 ## [3.1.0] - 2024-04-24
 ### Added
 - Added support for Ocean GKE APIs.